refactor: modularize engine
