+++ src/conditions.ts
@@ -1,0 +1,42 @@
+import type {
+	Actor,
+	Condition,
+	InCondition,
+	NotCondition,
+	ReferenceCondition,
+} from "./types";
+
+/**
+ * Compare a context value against the provided condition.
+ *
+ * The function recursively evaluates nested objects and
+ * supports negation, inclusion lists and references to
+ * the actor performing the check.
+ */
+export const matchCondition = (
+	value: unknown,
+	condition: Condition,
+	actor: Actor,
+): boolean => {
+	if (condition && typeof condition === "object" && !Array.isArray(condition)) {
+		if ("not" in condition)
+			return !matchCondition(value, (condition as NotCondition).not, actor);
+		if ("in" in condition)
+			return (condition as InCondition).in.includes(
+				value as string | number | boolean,
+			);
+		if ("reference" in condition) {
+			return (
+				value === actor?.[(condition as ReferenceCondition).reference.actor]
+			);
+		}
+		return Object.entries(condition).every(([k, c]) =>
+			matchCondition(
+				(value as Record<string, unknown>)?.[k],
+				c as Condition,
+				actor,
+			),
+		);
+	}
+	return value === condition;
+};
